requirements.txt
config==0.5.1
ipython==8.12.3
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.2.0
scikit_learn==1.2.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
utils==1.0.2
